+++ lib/metrics/cloudwatch.go
@@ -21,7 +21,6 @@ import (
 func init() {
 	Constructors[TypeCloudWatch] = TypeSpec{
 		constructor: NewCloudWatch,
-		Status:      docs.StatusBeta,
 		Summary: `
 Send metrics to AWS CloudWatch using the PutMetricData endpoint.`,
 		Description: `